Generating kubespec in case file is absent yet
When the rally pod is created, the /artifacts/kubespec_generated.yaml
file is not created yet, so the "bash /artifacts/rally-files/init-rally.sh"
cannot create the second ("kubernetes") env.
With this fix, if the kubespec file is absent, it is generated in /artifacts
directory where the certs and key files are also created.

Related-PROD: PROD-36128

Change-Id: Idf6840a602f2f2a88f09b6246d800640917f442a