Fixed dependencies around urllib pip module
Added a line to install 'urllib3<2.0' in the docker image.

The docker image fails to build due to the issue with the
dependencies. Firstly, 'urllib3<3,>=1.21.1' is installed
during the installation of the openstack packages.
After this, during installing the kubernetes pip module, it
requires 'google-auth>=1.0.1' which requires 'urllib3<2.0'
but the 2.0.3 version is already installed.

Related-PROD: PROD-37198
Change-Id: Ia238347528ee2f058e894460972996e26501651d