Added cleanup of free floating ip addresses
Deleting only free (not associated) FIPs which were
created inside the projects that can beleted (tempest, cvp,
etc).
Known issue: when run in -t dry-run mode, the servers are not
deleted yet, so the potential FIPs are still associated,
and the dry run log will not include them (if any).

Related-PROD: PROD-37187
Change-Id: I306d7bf63c11a7026ba5a8bd4ca9690541893911