Fixed the Dockerfile to download Ubuntu 20.04 packages
For the SPT and FIO tests we download the offline
deb packages for Ubuntu 20.04 cloud images and put
them to /opt/packages.

However the OS aptelope heat product image is based
on Ubuntu 22.04. So the packages for the 22.04 release
are downloaded instead of 20.04 which is used for FIO
and Iperf tests.

Thus, we download 20.04 deb packages firstly and then
build the image.

Related-PROD: PROD-37187
Change-Id: Iee4870a1b8d477d6cb17d8512456ceab12a8b112